backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
google-genai==0.3.0
googlemaps==4.10.0
black==24.4.2
flake8==7.1.0
